Styling tweaks after initial feedback
