+++ src/common/imageUrlBuilder.ts
@@ -15,3 +15,13 @@ export const imageUrlBuilder = unconfiguredImageUrlBuilder()
   .auto("format")
   .dpr(window.devicePixelRatio ?? 1)
   .quality(defaultQuality);
+
+export const getAspectRatio = (imageRef: string): number | undefined => {
+  const dimensionsArr = imageRef.match(/\d+x\d+/g);
+  if (!dimensionsArr) {
+    return undefined;
+  }
+  const dimensions = dimensionsArr.join().split("x");
+  const [width, height] = dimensions.map((n: string) => Number(n));
+  return width / height;
+};
